#367 Move the updater to their own classes
